organized, enemy sprite sheet. and movement bug fixed
